requirements.txt
langgraph==1.1.3
langchain-core==1.2.23
openai==2.30.0
anthropic==0.86.0
ollama==0.6.1
hypothesis==6.151.9
pytest==9.0.2
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.8.0
matplotlib==3.10.8
pyyaml==6.0.3
jsonschema==4.26.0
sentence-transformers==5.3.0
statsmodels==0.14.6
datasets==4.8.4
requests==2.33.0
